+++ packages/ignis/src/ignis/ignis_core.h
@@ -58,7 +58,7 @@ VkPhysicalDevice ignisGetVkPhysicalDevice();
 VkFormat ignisGetSwapchainImageFormat();
 VkFormat ignisGetSwapchainDepthFormat();
 
-uint32_t ignisGetMaxSamplerAnisotropy();
+float ignisGetMaxSamplerAnisotropy();
 
 uint32_t ignisGetCurrentFrame();
 uint32_t ignisGetQueueFamilyIndex(IgnisQueueFamily family);
